SVRC protobuf relay Issue SVO-1826

+++ WebsocketLibrary/Definition.h
@@ -1,7 +1,7 @@
 #pragma once
 namespace SvWsl
 {
-const unsigned Default_Port = 8080;
-const unsigned Default_SecondPort = 9000;
+const unsigned short Default_Port = 8080;
+const unsigned short Default_SecondPort = 9000;
 
 }
